Test/clik-mcp/search-service/requirements.txt
# Python Search Microservice for Council Minutes
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.10.0

# Search
bm25s==0.2.5
PyStemmer==2.2.0.3

# Embeddings & Vector Search
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
numpy>=1.26.0

# Korean text processing
konlpy==0.6.0

# LLM for contextual retrieval
google-genai==1.7.0
